Move handleApiDisplayResponse out of bl.cpp
Pure relocation of /api/display response flag handling (refresh rate,
special functions, status/filename, SF_REWIND / SF_SEND_TO_ME local image
paths) into display_session.

Export writeSpecialFunction for the session module. No intentional
behavior change; OG/X ifdefs and SF switch structure preserved.

Verified: pio run -e trmnl
